Move extension classes to Extensions folder
