+++ src/Equinor.ProCoSys.Completion.TieImport/Infrastructure/TILog.cs
@@ -1,4 +1,5 @@
-﻿using Statoil.TI.InterfaceServices.Message;
+﻿using Equinor.ProCoSys.Completion.TieImport.Extensions;
+using Statoil.TI.InterfaceServices.Message;
 
 namespace Equinor.ProCoSys.Completion.TieImport.Infrastructure;
